Move CompilationErrors to Errors project
