+++ CompilerTools/Errors/ErrorsStrategy.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Ivan Bondarev, Stanislav Mikhalkovich (for details please see \doc\copyright.txt)
 // This code is distributed under the GNU LGPL (for details please see \doc\license.txt)
-using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace PascalABCCompiler.Errors
 {
